Make retry logic more resilient to changes in xunit/vstest output
There was a significant change in output from dotnet test/vstest/xunit which caused us to bail too early and not retry properly.

We know the output (for now?) will contain the phrasing `Results file: {path}\{filename}.trx`, but we don't want to make that detection too fragile either. So we just check for the presence of the directory we pass as results dir, which should be always available in standard output.

Fixes #48