Fix broken verbose reporting of successful tests
We werre missing passing through the verbosity setting to the trx command, because we were creating a brand new setting class after the fact.

Since all the logic in that settings class (defaulting the path) is encapsulated in the Validate method, we can safely just use it when running the command.

For dotnet-retest we never do recursive log lookup, and we should preserve the legacy Skipped setting if it's provided too.